pyproject.toml
[project]
name = "image-search-service"
version = "0.1.0"
description = "Image search service with vector similarity"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "rq>=2.0.0",
    "qdrant-client>=1.12.0",
    "typer[all]>=0.15.0",
    "open-clip-torch>=2.24.0",
    "torch>=2.0.0",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.9",
    "watchdog>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "aiosqlite>=0.19.0",
]

[project.scripts]
image-search-cli = "image_search_service.scripts.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["rq.*", "qdrant_client.*", "open_clip.*", "torch.*", "PIL.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.0",
    "pillow>=12.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
